Crate i18n_loader

Re-exports

pub use lazy_static;

Macros

langid

langids

Structs

AttrCache

Provides a cache for the value of a specific attribute from a localization entry.

FluentArgs

Fluent messages can use arguments in order to programmatically add values to a translated string. For instance, in a localized application you may wish to display a user’s email count. This could be done with the following message.

FluentResource

A resource containing a list of localization messages.

LanguageIdentifier

LanguageIdentifier is a core struct representing a Unicode Language Identifier.

Locale

Manages Fluent localization resources for a specific locale.

Locales

A thread-safe container for all loaded localization data.

Message

Represents a localized message with its ID, value, and attributes.

Query

Represents a request to format a localized message, including its ID and arguments.

Enums

FluentError

Core error type for Fluent runtime system.

FluentValue

The FluentValue enum represents values which can be formatted to a String.

ReferenceKind

Maps an [InlineExpression] into the kind of reference, with owned strings that identify the expression. This makes it so that the [InlineExpression] can be used to generate an error string.

ResolverError

Errors generated during the process of resolving a fluent message into a string. This process takes place in the write method of the WriteValue trait.

Type Aliases

FluentBundle

Specialized FluentBundle over concurrent [IntlLangMemoizer].